🔨 refactor: 폴더 구조 리팩토링 및 ProductControl 컴포넌트 분리
