+++ src/pages/ItemsPage/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import TopNavigation from 'components/TopNavigation';
-import BestProductList from 'components/BestProductList';
-import AllProductList from 'components/AllProductList';
+import BestProductList from './components/BestProductList';
+import AllProductList from './components/AllProductList';
 import './style.css';
 
 const ItemsPage = () => {
